requirements.txt
aiogram==2.25.1
python-dotenv~=1.0.0
psycopg2==2.9.6
selenium==4.9.0
